Calendar - USA Federal

cal_usa_federal() is a federal calendar for the United States.

cal_usa_federal(since = "1900-01-01", until = "2100-01-01", observed = TRUE)

Arguments

since
[Date(1)]

The lower bound on where to begin looking for the holiday.

until
[Date(1)]

The upper bound on where to end looking for the holiday.

observed
[logical(1)]

If the holiday falls on a weekend, should the observed date be returned? The observed date is the nearest weekday. If FALSE, the exact holiday date is returned.

Value

An rcalendar.

Holidays

Martin Luther King Jr. Day

Christmas

New Year's Day

Examples

library(almanac)

cal <- cal_usa_federal()

alma_search("2000-01-01", "2001-01-01", cal)

#> [1] "2000-01-17" "2000-12-25" "2001-01-01"